+++ vcd_cli/vcd.py
@@ -132,6 +132,7 @@ else:
     from vcd_cli import task  # NOQA
     from vcd_cli import user  # NOQA
     from vcd_cli import vapp  # NOQA
+    from vcd_cli import vapp_network  # NOQA
     from vcd_cli import vc  # NOQA
     from vcd_cli import vdc  # NOQA
     from vcd_cli import vm  # NOQA
